opea-comps/mega-service-new/requirements.txt
opea-comps
fastapi
uvicorn
opentelemetry-api
opentelemetry-sdk
opentelemetry-instrumentation-fastapi
jinja2